third_party/WebKit/LayoutTests/fast/files/workers/worker-file-last-modified.html

<!DOCTYPE html>
<script src="../../../resources/js-test.js"></script>
<script>
description("Test that File.lastModifiedDate works in workers.");
startWorker('resources/worker-file-last-modified.js');
</script>
