third_party/blink/web_tests/fast/files/workers/worker-file-last-modified-expected.txt

Test that File.lastModifiedDate works in workers.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Starting worker: resources/worker-file-last-modified.js
PASS [Worker] file.lastModifiedDate.valueOf() is 1234567890
PASS successfullyParsed is true

TEST COMPLETE
